refactor : 스케줄 결과 분석 로직을 ScheduleService에서 ScheduleResultAnalysisService로 분리
